% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{ode_functions}
\alias{ode_functions}
\title{ode_functions}
\usage{
ode_functions(x_old, dt, mx = 10, N_end = 40, n = 4, N = 13, a = 4,
  a_s = 2.2, a_e = 6, k = 1, eta = 0.25, eta_m = 0.125,
  mature_mu = 0, eta_b = 0.1)
}
\arguments{
\item{x_old}{A named vector containing gene expression for 13 genes in the minimal network.}

\item{dt}{The time step used in the simulation}

\item{mx}{the number of lags to include}

\item{n}{the number of lags to include}

\item{N}{the number of lags to include}

\item{a}{the number of lags to include}

\item{a_s}{the number of lags to include}

\item{a_e}{the number of lags to include}

\item{k}{the number of lags to include}

\item{eta}{the number of lags to include}

\item{eta_m}{the number of lags to include}

\item{mature_mu}{the number of lags to include}

\item{eta_b}{the number of lags to include}
}
\value{
the predicted value of y
}
\description{
The set of ODE functions describing the minimal network for the central nervous system development.
}
\examples{
x_old <- c(2, rep(0, 12))
names(x_old) <- c("Pax6","Mash1","Brn2","Zic1","Tuj1","Hes5","Scl","Olig2","Stat3","Myt1L","Aldh1L","Sox8","Mature")

}
\seealso{
\code{\link{simulate_cns}}
}
\keyword{predict}
