% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{simulate_cns}
\alias{simulate_cns}
\title{simulate_cns}
\usage{
simulate_cns(steps = 1000, cells = 25, N_end = 40, mx = 10, n = 4,
  N = 13, a = 4, a_s = 2.2, a_e = 6, k = 1, eta = 0.25,
  eta_m = 0.125, mature_mu = 0, eta_b = 0.1)
}
\arguments{
\item{steps}{Number of total steps for the simulation.}

\item{cells}{Numer of cells you want to simulate.}

\item{N_end}{The end time point for your simulation.}

\item{mx}{parameter mx in the ODE model}

\item{n}{parameter mx in the ODE model}

\item{N}{parameter mx in the ODE model}

\item{a}{parameter mx in the ODE model}

\item{a_s}{parameter mx in the ODE model}

\item{a_e}{parameter mx in the ODE model}

\item{k}{parameter mx in the ODE model}

\item{eta}{parameter mx in the ODE model}

\item{eta_m}{parameter mx in the ODE model}

\item{mature_mu}{parameter mx in the ODE model}

\item{eta_b}{parameter mx in the ODE model}
}
\value{
the predicted value of y
}
\description{
This function simulates the differentiation process of central nervous system and generating a time series
where different cells differentiate into different cell lineages (neuron, astrocyte or oligodendrocytes).
}
\examples{
simulate_cns()

}
\seealso{
\code{\link{ode_functions}}
}
\keyword{predict}
